First draft of the Sign-up form project.
This is my first attempt at the Odin Project's Sign-up form project. I made it in dark mode
for personal aesthetics, but otherwise tried to follow to project template as closely
as possible. Page has limited functionality, but JS scripts will test if the passwords match
and display red borders/ a popup if they do not and green borders if they do.
Right now, if you click the login, it will first verify that all fields are correct and then
it will just change to bottom text to display a message with your first name.