First attempt at Odin Project Admin Dashboard Project
This is my first go at the Odin Project Admin Dashboard project, with a primary
focus on using grids. Overall, it doesn't look that much like the project,
but I'm honestly not sure how much longer I can mess around with margins and column starts
before I go crazy, so I'm calling it here for now. I changed some of the icons around
and added a little user interactivity (some of the elements will display shadows or
underline if the user hovers over them), but there is no javascript file with this.