Added TicTacToe project
This update includes a new file for the TicTacToe project. It produces a basic unstyled 3x3 grid and allows
two players to play the game, switching markers on each click. Currently not totally functional, as it relies on the console
to display victory and only allows for one game. More styling is needed, but I'm honeslty kinda burned out on this project so I
may but that on the backburner for now.